requirements.txt
asgiref==3.6.0
cfgv==3.3.1
channels==4.0.0
distlib==0.3.6
Django==4.1.5
django-extensions==3.2.1
filelock==3.9.0
graphql-core==3.2.3
identify==2.5.13
nodeenv==1.7.0
Pillow==9.4.0
platformdirs==2.6.2
pre-commit==2.21.0
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.7
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
strawberry-django-auth==0.374.0
strawberry-django-plus==1.35
strawberry-graphql==0.154.0
strawberry-graphql-django==0.9
typing_extensions==4.4.0
tzdata==2022.7
virtualenv==20.17.1
